Create a launcher on the desktop for Windows (.bat), Mac (.command), or Linux (.sh)

Create a launcher on the desktop for Windows (.bat), Mac (.command), or Linux (.sh)

launcher(
  app = c("radiant", "radiant.data", "radiant.design", "radiant.basics",
    "radiant.model", "radiant.multivariate"),
  port = 4444,
  pdir = Sys.getenv("HOME")
)

Arguments

app

App to run when the desktop icon is double-clicked ("analytics", "marketing", "quant", or "base"). Default is "analytics"

port

Port to use for shiny::runApp (e.g, port = 4444)

pdir

Project directory to use. Default is Sys.getenv("HOME")

Details

On Windows/Mac/Linux a file named radiant.bat/radiant.command/radiant.sh will be put on the desktop. Double-click the file to launch the specified Radiant app

See also

Examples

if (FALSE) {
radiant::launcher("radiant.model")
}